Add files via upload
Starfieldq1.py is an animated starfield script with the center of the field following mouse xy. It uses the same queue and thread techniques as before so can be added to the existing display. The Graphics.py GraphWin class is extendend to add mouse xy motion event binding to Tkinter.